+++ apps/example/app/api/admin/[[...nextadmin]]/route.ts
@@ -1,12 +1,12 @@
-import { createAppHandler } from "@premieroctet/next-admin/dist/handler";
-import schema from "@/prisma/json-schema/json-schema.json";
-import { prisma } from "@/prisma";
 import { options } from "@/options";
+import { prisma } from "@/prisma";
+import schema from "@/prisma/json-schema/json-schema.json";
+import { createHandler } from "@premieroctet/next-admin/dist/appHandler";
 
-const { run } = createAppHandler({
+const { run } = createHandler({
   options,
   prisma,
   schema,
 });
 
-export { run as POST, run as GET, run as DELETE };
+export { run as DELETE, run as GET, run as POST };
