+++ apps/example/pages/pagerouter/admin/[[...nextadmin]].tsx
@@ -1,5 +1,6 @@
 import { AdminComponentProps, NextAdmin } from "@premieroctet/next-admin";
-import { GetServerSideProps, GetServerSidePropsResult } from "next";
+import { getNextAdminProps } from "@premieroctet/next-admin/dist/pageRouter";
+import { GetServerSideProps } from "next";
 import { options } from "../../../pageRouterOptions";
 import { prisma } from "../../../prisma";
 import schema from "../../../prisma/json-schema/json-schema.json";
@@ -22,13 +23,5 @@ export default function Admin(props: AdminComponentProps) {
   );
 }
 
-export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
-  const { nextAdminRouter } = await import(
-    "@premieroctet/next-admin/dist/router"
-  );
-
-  const adminRouter = await nextAdminRouter(prisma, schema, pageOptions);
-  return adminRouter.run(req, res) as Promise<
-    GetServerSidePropsResult<{ [key: string]: any }>
-  >;
-};
+export const getServerSideProps: GetServerSideProps = async ({ req }) =>
+  await getNextAdminProps(prisma, schema, pageOptions, req);
