+++ apps/example/tailwind.config.js
@@ -146,5 +146,5 @@ module.exports = {
     },
   ],
   plugins: [require("@headlessui/tailwindcss")],
-  presets: [require("@premieroctet/next-admin/dist/preset")],
+  presets: [require("@premieroctet/next-admin/tailwind-preset")],
 };
