+++ packages/next-admin/src/appRouter.ts
@@ -1,9 +1,8 @@
 "use server";
 import { GetPropsFromParamsParams } from "./types";
-import {
-  getPropsFromParams as _getPropsFromParams,
-} from "./utils/props";
+import { getPropsFromParams as _getPropsFromParams } from "./utils/props";
 
-export const getPropsFromParams = (
-  params: Omit<GetPropsFromParamsParams, "isAppDir">
-) => _getPropsFromParams({ ...params, isAppDir: true });
+export const getNextAdminProps = ({
+  isAppDir = true,
+  ...params
+}: GetPropsFromParamsParams) => _getPropsFromParams({ ...params, isAppDir });
