+++ packages/next-admin/src/pageHandler.ts
@@ -46,7 +46,7 @@ type CreateAppHandlerParams<P extends string = "nextadmin"> = {
   schema: any;
 };
 
-export const createApiRouter = <P extends string = "nextadmin">({
+export const createHandler = <P extends string = "nextadmin">({
   options,
   prisma,
   paramKey = "nextadmin" as P,
