+++ packages/next-admin/src/utils/props.ts
@@ -40,7 +40,7 @@ export async function getPropsFromParams({
   options,
   schema,
   prisma,
-  isAppDir = false,
+  isAppDir = true,
   locale,
   getMessages,
 }: GetPropsFromParamsParams): Promise<
